# application/src/jreleaser/templates/article/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Ublhub API",
    "description" : "Ublhub REST API definitions for you to consume.",
    "termsOfService" : "https://github.com/project-openubl/ublhub/blob/master/CODE_OF_CONDUCT.md",
    "contact" : {
      "name" : "Ublhub API Support",
      "url" : "https://project-openubl.github.io/",
      "email" : "projectopenubl@googlegroups.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "2.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
  "paths" : {
...
    "/api/api/projects" : {
      "get" : {
        "tags" : [ "Project Resource" ],
        "summary" : "List projects",
        "description" : "List all projects",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        } ]
      },
      "post" : {
        "tags" : [ "Project Resource" ],
        "summary" : "Create project",
        "description" : "Create a project",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      }
    },
...
    "/api/api/projects/{projectId}/companies" : {
      "get" : {
        "tags" : [ "Company Resource" ],
        "summary" : "List companies",
        "description" : "List all companies",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CompanyDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        } ]
      },
      "post" : {
        "tags" : [ "Company Resource" ],
        "summary" : "Create company",
        "description" : "Create a company",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompanyDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompanyDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      }
    },
    "/api/api/projects/{projectId}/companies/{companyId}" : {
      "get" : {
        "tags" : [ "Company Resource" ],
        "summary" : "Get company",
        "description" : "Get one company",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompanyDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write", "project:read" ]
        } ]
      },
      "put" : {
        "tags" : [ "Company Resource" ],
        "summary" : "Update company",
        "description" : "Update one company",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompanyDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompanyDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      },
      "delete" : {
        "tags" : [ "Company Resource" ],
        "summary" : "Delete company",
        "description" : "Delete one company",
        "parameters" : [ {
          "name" : "companyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        }, {
          "SecurityScheme" : [ "admin:app", "project:write" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CompanyDto" : {
        "required" : [ "ruc", "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "ruc" : {
            "maxLength" : 11,
            "minLength" : 11,
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "maxLength" : 250,
            "type" : "string"
          },
          "sunat" : {
            "$ref" : "#/components/schemas/SunatDto"
          }
        }
      },
...
      "ProjectDto" : {
        "required" : [ "name", "sunat" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "maxLength" : 250,
            "type" : "string"
          },
          "sunat" : {
            "$ref" : "#/components/schemas/SunatDto"
          }
        }
      },
...
      "SunatDto" : {
        "required" : [ "facturaUrl", "guiaUrl", "retencionUrl", "username", "password" ],
        "type" : "object",
        "properties" : {
          "facturaUrl" : {
            "type" : "string"
          },
          "guiaUrl" : {
            "type" : "string"
          },
          "retencionUrl" : {
            "type" : "string"
          },
          "username" : {
            "maxLength" : 250,
            "minLength" : 3,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 250,
            "minLength" : 3,
            "type" : "string"
          }
        }
      },
...
    }
  }
}